almost fixed all bugs

+++ app/views/classes/index.php
@@ -7,6 +7,8 @@
         <tr>
             <td style="text-align: center; margin-bottom: 2rem;">
                 <?= $data['Teacher']->teacherFirstName . ' ' . $data['Teacher']->teacherLastName ?>
+                <br>
+                <?= "Woont op de " . $data['Teacher']->teacherAdres ?>
             </td>
         </tr>
     </tbody>
@@ -17,12 +19,10 @@
 <div class="float-right" style="margin-bottom: 2rem;">
     <a class="btn btn-success" href="<?= URLROOT; ?>/classesController/create/<?= $data['Teacher']->teacherId ?>">Create
         new Class</a>
-    <a class="btn btn-danger"
-        href="<?= URLROOT; ?>/subjectsController/deleteTeacher/<?= $data['Teacher']->teacherId . '+' . $data['educationId'] ?>">Delete
+    <a class="btn btn-danger" href="<?= URLROOT; ?>/subjectsController/deleteTeacher/<?= $data['Teacher']->teacherId  ?>">Delete
         teacher</a>
 
-    <a class="btn btn-info"
-        href="<?= URLROOT; ?>/subjectsController/updateTeacher/<?= $data['Teacher']->teacherId ?>">edit
+    <a class="btn btn-info" href="<?= URLROOT; ?>/subjectsController/updateTeacher/<?= $data['Teacher']->teacherId ?>">edit
         Teacher</a>
 
 </div>
@@ -37,20 +37,16 @@
     </thead>
     <tbody>
         <?php foreach ($data["Class"] as $class) { ?>
-        <tr>
+            <tr>
 
-            <td><a href="<?= URLROOT; ?>studentsController/index/<?= $class->classId ?>"
-                    style="text-decoration:none; color:black;"><?= $class->classId ?></a></td>
-            <td><?= $class->classTeacherId ?></td>
-            <td><?= $class->className ?></td>
-            <td style="align-items: center;">
-                <a class=" btn btn-danger"
-                    href="<?= URLROOT; ?>/classesController/delete/<?= $class->classId . '+' . $data['Teacher']->teacherId  ?>"
-                    onclick="return confirm('Are you sure you want to delete this school?')">Delete</a>
-                <a class="btn btn-info"
-                    href="<?= URLROOT; ?>/classesController/update/<?= $class->classId . '+' . $data['Teacher']->teacherId ?>">Update</a>
-            </td>
-        </tr>
+                <td><a href="<?= URLROOT; ?>studentsController/index/<?= $class->classId ?>" style="text-decoration:none; color:black;"><?= $class->classId ?></a></td>
+                <td><?= $class->classTeacherId ?></td>
+                <td><?= $class->className ?></td>
+                <td style="align-items: center;">
+                    <a class=" btn btn-danger" href="<?= URLROOT; ?>/classesController/delete/<?= $class->classId . '+' . $data['Teacher']->teacherId  ?>" onclick="return confirm('Are you sure you want to delete this school?')">Delete</a>
+                    <a class="btn btn-info" href="<?= URLROOT; ?>/classesController/update/<?= $class->classId . '+' . $data['Teacher']->teacherId ?>">Update</a>
+                </td>
+            </tr>
         <?php } ?>
     </tbody>
 </table>
